added login test
git-svn-id: https://svn.atomicobject.com/svn/systir/trunk@46 5e93408f-b7ed-0310-b26a-d648f8c3d419

+++ example/mondo_pizza/systest/run_test.rb
@@ -17,7 +17,7 @@ end
 if File.extname(testfile) == '.idea'
 	puts "Converting idea file to test file..."
 	outfile = testfile.gsub('.idea', '.test')
-  FileUtils.rm outfile
+  FileUtils.rm outfile if File.exist?(outfile)
 	converter = TestFileMarkup.new
 	
 	out = converter.convert(testfile)
